pyproject.toml
[tool.poetry]
name = "repl_nix_ChatGPT-Line-Bot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
youtube-transcript-api = "^0.6.0"
bs4 = "^0.0.1"
Flask = "^2.2.3"
line-bot-sdk = "^2.4.2"
OpenCC = "^1.1.6"
pymongo = "^4.3.3"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
pandas = "^2.0.1"
auto-mix-prep = "^0.2.0"
gspread = "^5.11.0"
oauth2client = "^4.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
